==> preprocessing/S2c_Fly_Cell_Atlas_HEAD_Oxphos_scoring_AUCell_68.ipynb <==
# Fly Cell Atlas - HEAD - AUCell Oxphos scoring
*This script is to be run before the Rmd script, for computing the AUC values with pySCENIC* 

**Author:** Vincent Gardeux

**Date Created:** 13/02/2025
**Date Last Modified:** 13/02/2025

# import dependencies
import pandas as pd
import numpy as np
import h5py
from ctxcore.genesig import GeneSignature
from pyscenic.aucell import create_rankings, enrichment

GENE_SIGNATURE_FNAME = './data/Oxphos_genes.xlsx'
GENE_SIGNATURE_SNAME = '68 OXPHOS genes'
EXPRESSION_MTX_FNAME = './data/s_fca_biohub_head_10x.loom' # From https://cloud.flycellatlas.org/index.php/s/jtH5fGwLqtiSRSD

data_excel = pd.read_excel(io = GENE_SIGNATURE_FNAME, sheet_name = GENE_SIGNATURE_SNAME)
data_excel

#gs = GeneSignature('Flybase - GO - Oxidative phosphorylation', data_excel["Gene_Ensembl_ID"].tolist())
gs = GeneSignature('Flybase - GO - Oxidative phosphorylation', data_excel["Gene_Ensembl_ID"].tolist())
gs

len(gs)

# Open Loom file in reading mode
f = h5py.File(EXPRESSION_MTX_FNAME, 'r')
m = f["/matrix"][:,:]
gene_names = f["/row_attrs/Gene"].asstr()[:]
cell_names = f["/col_attrs/CellID"].asstr()[:]
f.close()

m

m.shape

cell_names

len(cell_names)

gene_names

len(gene_names)

# [Input] Fixed gene annotation (for SCENIC feather file compatibility)
genome_data = pd.read_csv("./data/dmel-all-r6.31-premrna.gene_mapping.txt", sep="\t", na_filter=False)

# Update gene names using fixing table
gene_to_ensembl_dict = genome_data.set_index('gene_name')['gene_id'].to_dict()

# Check if all gene names are in the dictionary
missing_genes = [gene for gene in gene_names if gene not in gene_to_ensembl_dict]
if missing_genes:
    print(f"Error: The following genes are missing from the annotation table: {len(missing_genes)}")
else:
    # Fix gene names
    gene_names_updated = np.array([
        gene_to_ensembl_dict[gene] for gene in gene_names
    ])

    print(gene_names_updated)

len(gene_names_updated)

ex_matrix = pd.DataFrame(m, columns = cell_names, index = gene_names_updated)
ex_matrix

# Compute the intersection
intersection = list(set(gene_names_updated).intersection(set(data_excel["Gene_Ensembl_ID"].tolist())))
print(intersection)

len(intersection)

# Compute the genes that don't overlap in the other direction
difference_reverse = list(set(data_excel["Gene_Ensembl_ID"].tolist()).difference(set(gene_names_updated)))
print(difference_reverse)

ND-B8 (FBgn0040705) is missing from the Fly-Cell-Atlas dataset. It is in the mapping file, so most likely was filtered out.

# Run AUCell
# 1. Rankings (~2mn)
rnk_mtx = create_rankings(ex_matrix.transpose(), seed=42)

# 2. Enrichment (~5s)
aucs = enrichment(rnk_mtx, gs)
aucs

min(aucs["AUC"])

max(aucs["AUC"])

# Save DataFrame to TSV file
aucs.to_csv("./data/FCA_68_Oxphos_AUCell_auc.tsv", sep='\t', index=True)
