==> preprocessing/S2f_Martirosyan_PD_UPR_scoring_AUCell.ipynb <==
# GSE243639 - Martirosyan - AUCell UPR scoring
*This script is to be run before the Rmd script, for computing the AUC values with pySCENIC* 

**Author:** Vincent Gardeux

**Date Created:** 04/03/2025
**Date Last Modified:** 04/03/2025

# import dependencies
import pandas as pd
import polars as pl
import numpy as np
from ctxcore.genesig import GeneSignature
from pyscenic.aucell import create_rankings, enrichment

GENE_SIGNATURE_FNAME = '/data/gardeux/Neuro_Droso_ND75KD/data/UPR_genes_REACTOME_human.txt' # From https://www.gsea-msigdb.org/gsea/msigdb/human/geneset/REACTOME_UNFOLDED_PROTEIN_RESPONSE_UPR.html
GENE_SIGNATURE_SNAME = '93 UPR genes' # All are present in Martirosyan's data
EXPRESSION_MTX_FNAME = '/data/gardeux/Neuro_Droso_ND75KD/data/Martirosyan_2024/GSE243639_Filtered_count_table.csv.gz' # From https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE243639

data_genes = pd.read_csv(filepath_or_buffer=GENE_SIGNATURE_FNAME, header=None)[0].tolist()
gs = GeneSignature('REACTOME - UNFOLDED PROTEIN RESPONSE (UPR)', data_genes)
gs

len(gs)

# Open CSV count matrix
m = pl.read_csv(EXPRESSION_MTX_FNAME)

# Extract gene names (first column) and cell names (column headers)
gene_names = m[:, 0].to_list()  # First column as list of strings
cell_names = m.columns[1:]      # Column names (excluding first column)

m

m.shape

len(cell_names)

len(gene_names)

ex_matrix = pd.DataFrame(m[:, 1:], columns = cell_names, index = gene_names)
ex_matrix

# Compute the intersection
intersection = list(set(gene_names).intersection(set(data_genes)))
print(intersection)

len(intersection)

# Compute the genes that don't overlap in the other direction
difference_reverse = list(set(data_genes).difference(set(gene_names)))
print(difference_reverse)

No gene is missing from the dataset.

# Run AUCell
# 1. Rankings (~2mn)
rnk_mtx = create_rankings(ex_matrix.transpose(), seed=42)

# 2. Enrichment (~5s)
aucs = enrichment(rnk_mtx, gs)
aucs

min(aucs["AUC"])

max(aucs["AUC"])

# Save DataFrame to TSV file
aucs.to_csv("/data/gardeux/Neuro_Droso_ND75KD/data/Martirosyan_93_UPR_AUCell_auc.tsv", sep='\t', index=True)
