==> preprocessing/S2j_Jeffries_HBAging_UPR_scoring_AUCell.ipynb.ipynb <==
# Jeffries Human Brain Aging paper - AUCell UPR scoring
*This script is to be run before the Rmd script, for computing the AUC values with pySCENIC* 

**Author:** Vincent Gardeux

**Date Created:** 2025-11-17
**Date Last Modified:** 2025-11-17

# import dependencies
import pandas as pd
import numpy as np
import anndata as ad
import time
from ctxcore.genesig import GeneSignature
from pyscenic.aucell import create_rankings, enrichment

GENE_SIGNATURE_FNAME = '/data/gardeux/Neuro_Droso_ND75KD/data/UPR_genes_REACTOME_human.txt' # From https://www.gsea-msigdb.org/gsea/msigdb/human/geneset/REACTOME_UNFOLDED_PROTEIN_RESPONSE_UPR.html
GENE_SIGNATURE_SNAME = '90 + 3 UPR genes' # 3 are missing from Jeffries' data
EXPRESSION_H5AD_FNAME = '/data/gardeux/Neuro_Droso_ND75KD/data/Jeffries_2025_HumanBrainAging/pfc.clean.h5ad' # Built from pfc.clean.rds. Downloaded from https://publications.wenglab.org/SomaMut/Jeffries_Yu_BrainAging_2025/

data_genes = pd.read_csv(filepath_or_buffer=GENE_SIGNATURE_FNAME, header=None)[0].tolist()
gs = GeneSignature('REACTOME - UNFOLDED PROTEIN RESPONSE (UPR)', data_genes)
gs

len(gs)

# [Input] Load expression matrix from H5ad file
f_h5ad = ad.read_h5ad(EXPRESSION_H5AD_FNAME)
f_gene_names = f_h5ad.var_names.tolist()  # Gene names
f_cell_names = f_h5ad.obs_names.tolist()   # Cell names
ex_matrix = pd.DataFrame.sparse.from_spmatrix(f_h5ad.X.T, index=f_gene_names, columns=f_cell_names) # 33538 genes x 2663736 cells
ex_matrix

# Compute the intersection
intersection = list(set(f_gene_names).intersection(set(data_genes)))
print(intersection)

len(intersection)

# Compute the genes that don't overlap in the other direction
difference_reverse = list(set(data_genes).difference(set(f_gene_names)))
print(difference_reverse)

len(difference_reverse)

- TSPYL1 TSPYL4 TSPYL5 TSPYL6 are in the matrix. But TSPYL2 is not there.
- EIF2S1 EIF2S2 EIF2S3B are in the matrix. But EIF2S3 is not there.
- MBTPS1 is in the matrix. But MBTPS2 is not there.

So I assume their absence simply means that they are not expressed in the data. 3 genes are thus missing from the dataset (TSPYL2, EIF2S3, MBTPS2).

# Run AUCell
# 1. Rankings (~30mn)
start_time = time.time()

rnk_mtx = create_rankings(ex_matrix.transpose(), seed=42)

print(f"Ranking (AUCell) computing time: {(time.time() - start_time):.2f} seconds")

# 2. Enrichment (~5s)
aucs = enrichment(rnk_mtx, gs)
aucs

min(aucs["AUC"])

max(aucs["AUC"])

# Save DataFrame to TSV file
aucs.to_csv("/data/gardeux/Neuro_Droso_ND75KD/data/Jeffries_2025_HumanBrainAging_90_UPR_AUCell_auc.tsv", sep='\t', index=True)
