==> preprocessing/S3a_Pan_neuro_both_pySCENIC_pipeline_vEmbed_filtered.ipynb <==
# Emma Pan Neuro (Control + ND75KD) - pySCENIC pipeline (Embedded version)

**Author:** Vincent Gardeux

**Date Created:** 03/06/2024

# Libraries

# Fix OPENBLAS Warnings
import os
default_n_threads = 12
os.environ['OPENBLAS_NUM_THREADS'] = f"{default_n_threads}"
os.environ['MKL_NUM_THREADS'] = f"{default_n_threads}"
os.environ['OMP_NUM_THREADS'] = f"{default_n_threads}"

# import dependencies
import pandas as pd
import numpy as np
import h5py
import pickle
import pytz

from datetime import datetime
from arboreto.utils import load_tf_names
from arboreto.algo import grnboost2
from distributed import Client, LocalCluster

from ctxcore.rnkdb import FeatherRankingDatabase as RankingDatabase
from pyscenic.binarization import binarize
from pyscenic.utils import modules_from_adjacencies
from pyscenic.prune import prune2df, df2regulons
from pyscenic.aucell import aucell

# Parameters

# [Input] Fixed gene annotation (for SCENIC feather file compatibility)
genome_data = pd.read_csv("/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/fixed_gene_annotation.tsv", sep = "\t", na_filter=False)

# [Input] Loom file to use
f_loom_path_scenic = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/Pan_neuro_both_reannotated_GFP_curated_reintegrated.loom"
# Open Loom file in reading mode
f = h5py.File(f_loom_path_scenic, 'r')
f_m = f["/matrix"][:,:]
f_gene_names = f["/row_attrs/Gene"].asstr()[:]
f_ensembl_names = f["/row_attrs/Accession"].asstr()[:]
f_cell_names = f["/col_attrs/CellID"].asstr()[:]
f.close()

# Update gene names using fixing table
ensembl_to_gene_dict = genome_data.set_index('Ensembl')['Name'].to_dict()
f_gene_names_updated = np.array([
    ensembl_to_gene_dict[ensembl] if ensembl in ensembl_to_gene_dict else gene
    for gene, ensembl in zip(f_gene_names, f_ensembl_names)
])

# Create expression matrix
ex_matrix = pd.DataFrame(f_m, columns = f_cell_names, index = f_gene_names_updated)
ex_matrix = ex_matrix.loc[f_gene_names_updated, :]

# [Input] Transcription factors list (SCENIC step 1: GRNBoost2)
#f_tfs = "/data/genome/drosophila_melanogaster/cistopic_flybase_r6.02/allTFs_dmel.txt" # From pySCENIC github
f_tfs = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/FBgg0000745_TF_Flybase.filtered.txt" # From Flybase
# Derive list of Transcription Factors(TF)
tf_names = load_tf_names(f_tfs)

# [Output] Adjacency matrix (SCENIC step 1: GRNBoost2)
adj_matrix = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_adj.csv"

# [Input] Ranking databases (SCENIC step 2-3: cisTarget)
f_db_names = ["/data/genome/drosophila_melanogaster/cistopic_flybase_r6.02/mc_v10_clust/dm6_v10_clust.genes_vs_motifs.rankings.feather"]
dbs = [RankingDatabase(fname=f_name, name=os.path.basename(f_name)) for f_name in f_db_names]

# [Input] Motif databases (SCENIC step 2-3: cisTarget)
f_motif_path = "/data/genome/drosophila_melanogaster/cistopic_flybase_r6.02/mc_v10_clust/motifs-v10-nr.flybase-m0.00001-o0.0.tbl"

# [Output] Regulons (SCENIC step 2-3: cisTarget)
f_motifs_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_motifs.tsv"
f_modules_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_modules.tsv"
f_regulons_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_regulons.tsv"
f_regulons_aucell_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_regulons_aucell.tsv"
f_regulons_binarized_aucell_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_regulons_aucell_binarized.tsv"
f_regulons_binarization_thresholds_aucell_path = "/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/filtered_tf_flybase/Pan_neuro_both_reannotated_GFP_curated_reintegrated_regulons_aucell_binarization_thresholds.tsv"

# Restrict matrix to feather genes
ranking_feather = pd.read_feather(f_db_names[0])
overlap_values = ex_matrix.index[pd.Series(ex_matrix.index).isin(ranking_feather.columns)].unique()
ex_matrix = ex_matrix.loc[overlap_values, :]

ex_matrix

# SCENIC steps

## STEP 1: Gene regulatory network inference, and generation of co-expression modules

### 1.a. GRN inference using the GRNBoost2 algorithm

In the initial phase of the pySCENIC pipeline the single cell expression profiles are used to infer co-expression modules from.

Run GRNboost from arboreto to infer co-expression modules

The arboreto package is used for this phase of the pipeline.

*Output:* List of adjacencies between a TF and its targets.

Run GRNBoost2 algorithm

start = datetime.now(pytz.timezone('Europe/Paris'))
print("Start time:", start.strftime("%H:%M:%S"))

# Prepare the multithreading
cluster = LocalCluster(name='grn_call', dashboard_address=":12345", n_workers=default_n_threads, threads_per_worker=8)
client = Client(cluster)

# Here I run the function within the package (no CLI)
adjacencies = grnboost2(expression_data=ex_matrix.transpose(), tf_names=tf_names, seed=42, verbose=True, client_or_address=client)
            
# Shutting down cluster
client.close()
cluster.close()
    
end = datetime.now(pytz.timezone('Europe/Paris'))
print("End time:", end.strftime("%H:%M:%S"))
print("Running time:", (end - start))
# Note: ~1h40 with n_workers=12, threads_per_worker=[not set] => default to 16 threads
# Note: ~2h10 with n_workers=12, threads_per_worker=8
# Note: 53mn with n_workers=12, threads_per_worker=8 => With reduced matrix

Read in the adjacencies matrix

adjacencies.to_csv(adj_matrix, index=False, sep=',')
#adjacencies = pd.read_csv(adj_matrix, sep=',', na_filter=False) # If na_filter=True, the nan gene is detected as NaN
adjacencies

## STEP 2-3: Regulon prediction aka cisTarget

*Output:* List of adjacencies between a TF and its targets.

### 2.a. Running regulon prediction using cisTarget

Here, we use the --mask_dropouts option, which affects how the correlation between TF and target genes is calculated during module creation. It is important to note that prior to pySCENIC v0.9.18, the default behavior was to mask dropouts, while in v0.9.18 and later, the correlation is performed using the entire set of cells (including those with zero expression). When using the modules_from_adjacencies function directly in python instead of via the command line, the rho_mask_dropouts option can be used to control this.

**Note:** Here I tried with and without the `rho_mask_dropouts=True` option. Without it, it generates 46 regulons. With it, 81 regulons. So I keep it.

**Update:** Now 99 regulons with fixed gene names, and `rho_mask_dropouts=True` option

**Update 2:** Now 108 regulons with fixed gene names, restricted to feather genes, and `rho_mask_dropouts=True` option. It contains dati and crc, but not repo anymore :(

**Update 3:** Now 764 regulons with filter_for_annotation=False ( no filtering on pruning )

**Update 4:** Now 579 regulons when using TF list from Flybase. 110 with basic pruning, and 289 with relaxed pruning (misses pros and sima).

modules = list(modules_from_adjacencies(adjacencies, ex_matrix.transpose(), rho_mask_dropouts=True, keep_only_activating=True)) # rho_mask_dropouts=True

modules_df = pd.DataFrame(index = range(0, len(modules)), columns = ("Regulon", "TF", "TFTargetGenesCorrelation", "NbMarkers", "Context", "NES", "Markers"))
for j in range(0, len(modules)):
    # Setting values
    context = list(modules[j].context)
    modules_df["Regulon"].iloc[j] = modules[j].name
    modules_df["TF"].iloc[j] = modules[j].transcription_factor
    modules_df["TFTargetGenesCorrelation"].iloc[j] = context[0]
    modules_df["NbMarkers"].iloc[j] = len(set(modules[j].gene2weight))
    modules_df["Context"].iloc[j] = context[1]
    modules_df["NES"].iloc[j] = modules[j].score
    modules_df["Markers"].iloc[j] = ','.join(list(modules[j].gene2weight))

modules_df = modules_df.sort_values(by='NbMarkers', ascending=False)
modules_df.to_csv(f_modules_path, index=False, sep = "\t")
modules_df

print(modules_df.TF.nunique(), "unique TF-modules were found ( out of",len(tf_names),"). Modules with less than 20 markers were filtered out.")

start = datetime.now(pytz.timezone('Europe/Paris'))
print("Start time:", start.strftime("%H:%M:%S"))

df = prune2df(dbs, modules, f_motif_path, num_workers=default_n_threads, weighted_recovery=True, rank_threshold = 1500, nes_threshold=1, motif_similarity_fdr=0.001, auc_threshold=0.05, filter_for_annotation=False)
    
end = datetime.now(pytz.timezone('Europe/Paris'))
print("End time:", end.strftime("%H:%M:%S"))
print("Running time:", (end - start))
# Note: 5mn37 with num_workers=48
# Note: 5mn29 with num_workers=12
# Note: 5mn33 with num_workers=10
# Note: 5mn37 with num_workers=8 ** optimal? 12mn with filter_for_annotation=False)
# Note: 8mn40 with num_workers=6
# Note: 9mn53 with num_workers=4
# Note: 30mn24 with num_workers=1

df.to_csv(f_motifs_path,sep = "\t")
#df

print(len(set(df.index.get_level_values('TF').values)), "regulons were kept, after pruning")

# Look for main regulons
print("pros", "pros" in df.index.get_level_values('TF').values, sep="\t")
print("dati", "dati" in df.index.get_level_values('TF').values, sep="\t")
print("scro", "scro" in df.index.get_level_values('TF').values, sep="\t")
print("crc", "crc" in df.index.get_level_values('TF').values, sep="\t")
print("repo", "repo" in df.index.get_level_values('TF').values, sep="\t")
print("sima", "sima" in df.index.get_level_values('TF').values, sep="\t")

print("Size of Dataframe:", len(df))
drop_indexes = []
for j in range(0, len(df)):
    # Setting values
    if(len(df["Enrichment"]["TargetGenes"][j]) == 0): drop_indexes.append(df.index[j])
df_filtered = df["Enrichment"].drop(index=drop_indexes)
print("Size of Dataframe:", len(df_filtered))

These "modules" are then combined into regulons, by taking the top NES for each TF (for main Motif, and final score of regulon). All genes are bundled together.

# This dataframe can then be converted to regulons.
regulons = df2regulons(df_filtered)

regulon_df = pd.DataFrame(index = range(0, len(regulons)), columns = ("Regulon", "TF", "TFTargetGenesCorrelation", "NbMarkers", "Motif", "NES", "Markers"))
for j in range(0, len(regulons)):
    # Fixing order of set
    context = list(regulons[j].context)
    if(context[0].endswith(".png")):
        tmp = context[0]
        context[0] = context[1]
        context[1] = tmp
    # Setting values
    regulon_df["Regulon"].iloc[j] = regulons[j].name
    regulon_df["TF"].iloc[j] = regulons[j].transcription_factor
    regulon_df["TFTargetGenesCorrelation"].iloc[j] = context[0]
    regulon_df["NbMarkers"].iloc[j] = len(set(regulons[j].gene2weight))
    regulon_df["Motif"].iloc[j] = "https://resources.aertslab.org/cistarget/motif_collections/v10nr_clust_public/logos/" + context[1]
    regulon_df["NES"].iloc[j] = regulons[j].score
    regulon_df["Markers"].iloc[j] = ','.join(list(regulons[j].gene2weight))

regulon_df = regulon_df.sort_values(by='NbMarkers', ascending=False)
regulon_df.to_csv(f_regulons_path, index=False, sep = "\t")
regulon_df

# Look for main regulons
print("pros", "pros" in regulon_df["TF"].values, sep="\t")
print("dati", "dati" in regulon_df["TF"].values, sep="\t")
print("scro", "scro" in regulon_df["TF"].values, sep="\t")
print("crc", "crc" in regulon_df["TF"].values, sep="\t")
print("repo", "repo" in regulon_df["TF"].values, sep="\t")
print("sima", "sima" in regulon_df["TF"].values, sep="\t")

## Phase III: Cellular regulon enrichment matrix (aka AUCell)

Characterize the different cells in a single-cell transcriptomics experiment by the enrichment of the regulons. Enrichment of a regulon is measures as AUC of the recovery curve of the genes that define this regulon.

auc_mtx = aucell(ex_matrix.transpose(), regulons, num_workers=default_n_threads)
auc_mtx.to_csv(f_regulons_aucell_path, sep = "\t")
auc_mtx

# Checkpoint to regenerate the object from the file
#auc_mtx = pd.read_csv("/home/gardeux/updepla_storage/users/gardeux/2023-04-19_snRNAseq_Emma/analysis/old/Pan_neuro_both_reannotated_GFP_curated_reintegrated_regulons_aucell.tsv", sep = "\t", index_col = "Cell")
auc_mtx = pd.read_csv(f_regulons_aucell_path, sep = "\t", index_col = "Cell")
auc_mtx.columns.name = "Regulon"
auc_mtx

start = datetime.now(pytz.timezone('Europe/Paris'))
print("Start time:", start.strftime("%H:%M:%S"))

auc_mtx_bin = binarize(auc_mtx, seed = 42, num_workers=default_n_threads)

end = datetime.now(pytz.timezone('Europe/Paris'))
print("End time:", end.strftime("%H:%M:%S"))
print("Running time:", (end - start))

# Note: 23mn12 with num_workers=12

binarization_thresholds = auc_mtx_bin[1]
binarization_thresholds.to_csv(f_regulons_binarization_thresholds_aucell_path, sep = "\t")
binarization_thresholds

auc_mtx_bin = auc_mtx_bin[0]
auc_mtx_bin.to_csv(f_regulons_binarized_aucell_path, sep = "\t")
auc_mtx_bin

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['scro(+)'] # 733 original (764 regulons, Aerts). 716 ok

binarization_thresholds.loc['scro(+)'] # 0.3359530388214861 original (764 regulons, Aerts). 0.2968568380139414

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['sima(+)'] # 1067 original (764 regulons, Aerts). 1212 ok

binarization_thresholds.loc['sima(+)'] # 0.21179422980092633 original (764 regulons, Aerts). 0.18676990882955136

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['crc(+)'] # 1136 original (764 regulons, Aerts). 1036 a bit too many?

binarization_thresholds.loc['crc(+)'] # 0.21042246303356155 original (764 regulons, Aerts). 0.23696722508584317

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['dati(+)'] # 11733 original (764 regulons, Aerts). 11631 ok

binarization_thresholds.loc['dati(+)'] # 0.25299783006969817 original (764 regulons, Aerts). 0.2806370838941694

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['pros(+)'] # 18482 original (764 regulons, Aerts). 18548 ok

binarization_thresholds.loc['pros(+)'] # 0.11976533052607628 original (764 regulons, Aerts). 0.13361937192963938

auc_mtx_bin.sum(numeric_only=True, axis=0).loc['repo(+)'] # 906 original (764 regulons, Aerts). 920 ok

binarization_thresholds.loc['repo(+)'] # 0.10907608108082668 original (764 regulons, Aerts). 0.11348899982841607

sum(auc_mtx["crc(+)"] > 0.238)

sum(auc_mtx["crc(+)"] > binarization_thresholds.loc['crc(+)'])

sum(auc_mtx_bin["crc(+)"])
