==> preprocessing/S4_Pan_neuro_integrated_add_metadata.Rmd <==
---
title: "Emma Pan Neuro (Control + ND75KD) - Generating final Seurat object (Oxphos AUCell, SCENIC scoring (continuous and binarized), and annotation"
author: "Vincent Gardeux"
date: "2024/03/13"
last_modified: "2024/10/25"
output:
  html_document:
    df_print: paged
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = "/data/gardeux/Neuro_Droso_ND75KD/")
```

## Libraries & functions

First, I'm loading the required libraries & functions

```{r}
suppressPackageStartupMessages(library(Seurat, lib.loc = "/software/Seuratv4/lib/")) # Seurat v4.4.0, library for single-cell analysis
suppressPackageStartupMessages(library(SeuratObject, lib.loc = "/software/Seuratv4/lib/")) # For compatibility purpose
suppressPackageStartupMessages(library(data.table)) # For writing DE gene file
suppressPackageStartupMessages(library(loomR)) # For building a file for ASAP
suppressPackageStartupMessages(library(Matrix)) # For building a file for ASAP
suppressPackageStartupMessages(library(ggplot2)) # For plotting
suppressPackageStartupMessages(library(ggpubr)) # For grid plotting
suppressPackageStartupMessages(library(plotly)) # For interactive plots
suppressPackageStartupMessages(library(crayon)) # Just for bolding the console output :D

cat(bold("Seurat"), "version", as.character(packageVersion("Seurat")), "\n")
cat(bold("SeuratObject"), "version", as.character(packageVersion("SeuratObject")), "\n")
cat(bold("data.table"), "version", as.character(packageVersion("data.table")), "\n")
cat(bold("loomR"), "version", as.character(packageVersion("loomR")), "\n")
cat(bold("Matrix"), "version", as.character(packageVersion("Matrix")), "\n")
cat(bold("ggplot2"), "version", as.character(packageVersion("ggplot2")), "\n")
cat(bold("ggpubr"), "version", as.character(packageVersion("ggpubr")), "\n")
cat(bold("plotly"), "version", as.character(packageVersion("plotly")), "\n")

# Color-blind friendly palette 3-colors
cbPalette <- c("#E69F00", "#000000", "#56B4E9")
steinPalette <- c(`Central Brain A`="#80C980",`Central Brain B`="#BDADD4",`Optic lobe`="#376CB0",`Glia`="#FBBF85",`Other`="#F0027E")

# Random seed
set.seed(42)
```

## Parameters

```{r}
input_seurat_integrated <- "./data/Pan_neuro_integrated.rds"
input_curated_annotations <- "./data/Pan_neuro_integrated_markers_annotation.txt"
input_SCENIC_regulons_path <- "./data/Pan_neuro_integrated_regulons_aucell.tsv"
input_SCENIC_regulons_binarized_path <- "./data/Pan_neuro_integrated_regulons_aucell_binarized.tsv"
input_Oxphos_117_AUCell_values_path <- "./data/Pan_neuro_integrated_117_Oxphos_AUCell_auc.tsv"
input_Oxphos_68_AUCell_values_path <- "./data/Pan_neuro_integrated_68_Oxphos_AUCell_auc.tsv"

output_seurat_final <- "./data/Pan_neuro_integrated_FINAL.rds"
```


## I. Reading Seurat object

First step is to read the Seurat objects, previously created

```{r}
message("Loading Seurat object...")
data.seurat <- readRDS(input_seurat_integrated)
message(ncol(data.seurat), " cells were loaded")
message(nrow(data.seurat), " genes were loaded")
```

## II.1 Reading curated annotation from Emma

```{r}
data.annotations <- fread(input_curated_annotations, sep = "\t", quote = F, header = T, data.table = F)
data.annotations
```

## II.2 Adding to Seurat object

```{r}
# Keep annotation-only, with "unannotated" when unknown
data.seurat@meta.data[["annotation"]] <- "unannotated"
for(cluster in data.annotations$cluster){
  data.seurat@meta.data[["annotation"]][data.seurat@meta.data[["seurat_clusters"]] == cluster] <- data.annotations[data.annotations$cluster == cluster, "annotation"]
}

# Adding cluster number to "unannotated" annotations
data.seurat$annotation_clusters <- data.seurat$annotation
data.seurat$annotation_clusters[data.seurat$annotation_clusters == "unannotated"] <- as.character(data.seurat$seurat_clusters)[data.seurat$annotation_clusters == "unannotated"]
```

## III.1 Reading regulons values from SCENIC output

```{r}
data.scenic <- fread(input_SCENIC_regulons_path, data.table = F)
rownames(data.scenic) <- data.scenic$Cell
data.scenic <- data.scenic[,-1]
data.scenic
```

Updating colnames

```{r}
colnames(data.scenic) <- gsub(x = colnames(data.scenic), pattern = "\\(\\+\\)", replacement = "_activating_regulon")
colnames(data.scenic) <- gsub(x = colnames(data.scenic), pattern = "\\(\\-\\)", replacement = "_repressing_regulon")
colnames(data.scenic) <- paste0("SCENIC_", colnames(data.scenic))
data.scenic
```

## III.2 Adding each regulon as a metadata to the Seurat object:

```{r}
# Adding each column as a metadata
for(regulon in colnames(data.scenic)){
  data.seurat@meta.data[[regulon]] <- data.scenic[colnames(data.seurat), regulon]
}
```

## III.3 Compute UMAP on regulon matrix

```{r fig.height=8, fig.width=12}
# Adding whole binarized regulon matrix as an embedding
colnames(data.scenic) <- paste0("SCENIC_", 1:ncol(data.scenic))
data.seurat@reductions[['SCENIC']] <- CreateDimReducObject(embeddings = as.matrix(data.scenic), key = 'SCENIC_', assay = 'RNA')

# Compute UMAP on it
data.seurat <- RunUMAP(object = data.seurat, reduction.name = "SCENICumap", reduction = "SCENIC", dims = 1:ncol(data.scenic), reduction.key = "SCENICumap_", assay = 'RNA', n.neighbors = 10L, metric = "correlation", min.dist = 0.4)
DimPlot(data.seurat, reduction = "SCENICumap", label = TRUE, pt.size = 0.5, group.by = "annotation") + NoLegend()
```

## III.4 Compute t-SNE on regulon matrix

```{r fig.height=8, fig.width=12}
# Compute t-SNE on it
data.seurat <- RunTSNE(object = data.seurat, reduction.name = "SCENICtsne", reduction = "SCENIC", dims = 1:ncol(data.scenic), reduction.key = "SCENICtsne_", assay = 'RNA')
DimPlot(data.seurat, reduction = "SCENICtsne", label = TRUE, pt.size = 0.5, group.by = "annotation") + NoLegend()
```

## IV.1 Reading regulons values from binarized SCENIC output

```{r}
data.scenic <- fread(input_SCENIC_regulons_binarized_path, data.table = F)
rownames(data.scenic) <- data.scenic$Cell
data.scenic <- data.scenic[,-1]
data.scenic
```

Updating colnames

```{r}
colnames(data.scenic) <- gsub(x = colnames(data.scenic), pattern = "\\(\\+\\)", replacement = "_activating_regulon_binarized")
colnames(data.scenic) <- gsub(x = colnames(data.scenic), pattern = "\\(\\-\\)", replacement = "_repressing_regulon_binarized")
colnames(data.scenic) <- paste0("SCENIC_", colnames(data.scenic))
data.scenic
```

## IV.2 Adding each regulon as a metadata to the Seurat object:

```{r}
# Adding each column as a metadata
for(regulon in colnames(data.scenic)){
  data.seurat@meta.data[[regulon]] <- data.scenic[colnames(data.seurat), regulon]
}
```

## IV.3 Compute UMAP on regulon matrix

```{r fig.height=8, fig.width=12}
# Adding whole binarized regulon matrix as an embedding
colnames(data.scenic) <- paste0("SCENIC_", 1:ncol(data.scenic))
data.seurat@reductions[['SCENICbinarized']] <- CreateDimReducObject(embeddings = as.matrix(data.scenic), key = 'SCENICbinarized_', assay = 'RNA')

# Compute UMAP on it
data.seurat <- RunUMAP(object = data.seurat, reduction.name = "SCENICbinarizedumap", reduction = "SCENICbinarized", dims = 1:ncol(data.scenic), reduction.key = "SCENICbinarizedumap_", assay = 'RNA', n.neighbors = 10L, metric = "correlation", min.dist = 0.4)
DimPlot(data.seurat, reduction = "SCENICbinarizedumap", label = TRUE, pt.size = 0.5, group.by = "annotation") + NoLegend()
```

## IV.4 Compute t-SNE on regulon matrix

```{r fig.height=8, fig.width=12}
# Compute t-SNE on it
data.seurat <- RunTSNE(object = data.seurat, reduction.name = "SCENICbinarizedtsne", reduction = "SCENICbinarized", dims = 1:ncol(data.scenic), reduction.key = "SCENICbinarizedtsne_", assay = 'RNA')
DimPlot(data.seurat, reduction = "SCENICbinarizedtsne", label = TRUE, pt.size = 0.5, group.by = "annotation") + NoLegend()
```

## Clean RAM

```{r}
rm(data.scenic)
rm(data.annotations)
```

## V. Predicting groups

```{r fig.width=18, fig.height=24}
plot.list <- list()
for(cluster in levels(data.seurat$seurat_clusters)){
  data.tmp <- as.data.frame(t(data.seurat@assays$RNA@data[c("Imp", "pros", "dati", "scro"), data.seurat$seurat_clusters == cluster]))
  data.tmp$pros_reg <- data.seurat@meta.data[rownames(data.tmp), "SCENIC_pros_activating_regulon"]
  data.tmp$dati_reg <- data.seurat@meta.data[rownames(data.tmp), "SCENIC_dati_activating_regulon"]
  data.tmp$scro_reg <- data.seurat@meta.data[rownames(data.tmp), "SCENIC_scro_activating_regulon"]
  data.tmp <- reshape2::melt(data.tmp, measure.vars = c("Imp", "pros", "pros_reg", "dati", "dati_reg", "scro", "scro_reg"), value.name = "expression", variable.name = "gene")
  plot.list[[cluster]] <- ggplot(data.tmp, aes(x = gene, y = expression)) + geom_boxplot() + ggtitle(paste0("Cluster ", cluster)) + theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1))
}
ggarrange(plotlist = plot.list, ncol = 9, nrow = 10)
```

## V.1 Genes or regulons?

```{r}
# First, build non-zero distribution for each marker gene
global.imp <- data.seurat@assays$RNA@data["Imp",]
global.imp <- global.imp[global.imp != 0]
hist(global.imp)
global.pros <- data.seurat@assays$RNA@data["pros",]
global.pros <- global.pros[global.pros != 0]
hist(global.pros)
global.dati <- data.seurat@assays$RNA@data["dati",]
global.dati <- global.dati[global.dati != 0]
hist(global.dati)
global.scro <- data.seurat@assays$RNA@data["scro",]
global.scro <- global.scro[global.scro != 0]
hist(global.scro)
global.repo <- data.seurat@assays$RNA@data["repo",]
global.repo <- global.repo[global.repo != 0]
hist(global.repo)
```

## V.2 Gene-based (not kept)

All gene distribution look normal-ish. But it's not the case for the regulon distribution.
If we work with gene-expression only, then for each cell, I would compute the p-values of belonging to the corresponding normal distribution.

```{r}
label <- rep("Ambiguous", ncol(data.seurat))
mn_imp <- mean(global.imp); sd_imp <- sd(global.imp); mn_pros <- mean(global.pros); sd_pros <- sd(global.pros); mn_dati <- mean(global.dati); sd_dati <- sd(global.dati); mn_scro <- mean(global.scro); sd_scro <- sd(global.scro); mn_repo <- mean(global.repo); sd_repo <- sd(global.repo)
imp_values <- data.seurat@assays$RNA@data["Imp",]; pros_values <- data.seurat@assays$RNA@data["pros",]; dati_values <- data.seurat@assays$RNA@data["dati",]; scro_values <- data.seurat@assays$RNA@data["scro",]; repo_values <- data.seurat@assays$RNA@data["repo",]
for(cell in 1:ncol(data.seurat)){
  p_imp <- pnorm(q = imp_values[cell], mean = mn_imp, sd = sd_imp)
  p_pros <- pnorm(q = pros_values[cell], mean = mn_pros, sd = sd_pros)
  p_dati <- pnorm(q = dati_values[cell], mean = mn_dati, sd = sd_dati)
  p_scro <- pnorm(q = scro_values[cell], mean = mn_scro, sd = sd_scro)
  p_repo <- pnorm(q = repo_values[cell], mean = mn_repo, sd = sd_repo)
  if(max(p_imp, p_pros, p_dati, p_scro, p_repo) == p_scro){
    label[cell] <- "Optic lobe"
  } else if(max(p_imp, p_pros, p_dati, p_scro, p_repo) == p_repo){
    label[cell] <- "Glia"
  } else if(max(p_imp, p_pros, p_dati, p_scro) == p_imp){
    label[cell] <- "Central Brain B"
  } else label[cell] <- "Central Brain A"
}
data.seurat$Neuron_category <- factor(label, levels = c("Central Brain A", "Central Brain B", "Optic lobe", "Glia", "Other"))
table(data.seurat$Neuron_category)
```

```{r fig.height=7, fig.width=9}
DimPlot(data.seurat, reduction = "tsne_harmony", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

```{r fig.height=6, fig.width=9}
DimPlot(data.seurat, reduction = "SCENICbinarizedumap", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

```{r fig.height=6, fig.width=9}
DimPlot(data.seurat, reduction = "SCENICbinarizedtsne", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

## V.3 Regulon-based (kept)

All gene distribution look normal-ish. But regulons are binarized.

```{r}
label <- rep("Ambiguous", ncol(data.seurat))
mn_imp <- mean(global.imp)
sd_imp <- sd(global.imp)
imp_values <- data.seurat@assays$RNA@data["Imp",]
for(cell in 1:ncol(data.seurat)){
  p_imp <- pnorm(q = imp_values[cell], mean = mn_imp, sd = sd_imp)
  if(data.seurat@meta.data$SCENIC_scro_activating_regulon_binarized[cell] == 1){
    label[cell] <- "Optic lobe"
  } else if(data.seurat@meta.data$SCENIC_repo_activating_regulon_binarized[cell] == 1){
    label[cell] <- "Glia"
  } else if(data.seurat@meta.data$SCENIC_dati_activating_regulon_binarized[cell] == 1 & data.seurat@meta.data$SCENIC_pros_activating_regulon_binarized[cell] == 1){
    label[cell] <- "Central Brain A"
  } else label[cell] <- "Central Brain B"
}
data.seurat$Neuron_category <- factor(label, levels = c("Central Brain A", "Central Brain B", "Optic lobe", "Glia", "Other"))
table(data.seurat$Neuron_category)
```

```{r fig.height=7, fig.width=9}
DimPlot(data.seurat, reduction = "tsne_harmony", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

```{r fig.height=6, fig.width=9}
DimPlot(data.seurat, reduction = "SCENICbinarizedumap", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

```{r fig.height=6, fig.width=9}
DimPlot(data.seurat, reduction = "SCENICbinarizedtsne", pt.size = 0.5, group.by = "Neuron_category", cols = steinPalette)
```

### Check Imp expression. It should be higher in Central Brain B

```{r}
# Prepare data
imp_data <- data.frame(
  Imp = as.vector(data.seurat@assays$RNA@data["Imp", ]),
  Neuron_category = data.seurat$Neuron_category
)

# Plot
ggplot(imp_data, aes(x = Neuron_category, y = Imp, fill = Neuron_category)) +
  geom_boxplot() +
  stat_summary(fun = "mean", geom = "point", shape = 18, size = 3, color = "black", fill = "white") + # Display mean points
  labs(
    title = "Imp expression by Neuron category",
    x = NULL,
    y = "Imp normalized expression"
  ) +
  scale_fill_manual(values = steinPalette) +  # Apply color palette
  theme_minimal() +
  stat_compare_means(
    method = "wilcox",
    comparisons = list(
      c("Central Brain A", "Central Brain B"), 
      c("Central Brain A", "Optic lobe"), 
      c("Central Brain A", "Glia"), 
      c("Central Brain B", "Optic lobe"), 
      c("Central Brain B", "Glia"), 
      c("Optic lobe", "Glia")
    ),
    label = "p.signif",
    label.x = 1.5
  )
```


## VI.1 Reading OXPHOS AUC values from AUCell (using the 117 OXPHOS genes from Flybase GO-BP GO:0006119 : oxidative phosphorylation)

```{r}
data.auc <- fread(input_Oxphos_117_AUCell_values_path, data.table = F)
rownames(data.auc) <- data.auc$Cell
data.auc
```

## VI.2 Adding to Seurat object

```{r}
data.seurat$AUC_OXPHOS_117_Flybase <- data.auc[colnames(data.seurat), "AUC"]
hist(data.seurat$AUC_OXPHOS_117_Flybase)
```

## VII.1 Reading OXPHOS AUC values from AUCell (using the 68 OXPHOS genes from Porcelli et al. https://link.springer.com/article/10.1186/1471-2148-7-215)

```{r}
data.auc <- fread(input_Oxphos_68_AUCell_values_path, data.table = F)
rownames(data.auc) <- data.auc$Cell
data.auc
```

## VII.2 Adding to Seurat object

```{r}
data.seurat$AUC_OXPHOS_68_Porcelli <- data.auc[colnames(data.seurat), "AUC"]
hist(data.seurat$AUC_OXPHOS_68_Porcelli)
```

## VIII. Save Seurat object

```{r}
saveRDS(data.seurat, file = output_seurat_final)
```
